added commandline runner to application to test the persistence of data to the datebase.

+++ src/main/resources/schema.sql
@@ -1,6 +1,7 @@
 
+
 create table if not exists exercise (
-    id int NOT NULL,
+    id BIGSERIAL NOT NULL,
     exercise_name VARCHAR(50) NOT NULL,
     type VARCHAR(50) NOT NULL,
     muscle VARCHAR(50) NOT NULL,
